backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
groq==0.11.0
sentence-transformers==3.1.1
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
resend==2.4.0
reportlab==4.2.2
Pillow==10.4.0
python-jose==3.3.0
apscheduler==3.10.4
